backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.3.6
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
rich==13.9.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
gunicorn==23.0.0
celery[redis]==5.5.3
flask_caching==2.3.1
alembic==1.14.0
flask-migrate==4.0.7

# Visual Processing Dependencies
playwright==1.40.0
Pillow==10.1.0
opencv-python==4.8.1.78
colorthief==0.2.1
webcolors==1.13
beautifulsoup4==4.12.2
lxml==4.9.3

# Presentation Generation Dependencies
reportlab==4.0.9
python-pptx==0.6.23
matplotlib==3.8.2
seaborn==0.13.0

# Async Processing Dependencies
aiohttp==3.9.1
aiofiles==23.2.0
asyncio-throttle==1.0.2
aiocache==0.12.2